chore: squash bridge config
