+++ src/bridges/ethereum-darwinia/config/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from './bridge';
