+++ src/bridges/ethereum-darwiniaDVM/config/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from './bridge';
