+++ src/bridges/substrate-dvm/config/bridge.ts
@@ -1,0 +1,38 @@
+import { omit } from 'lodash';
+import { pangolinConfig } from '../../../config/network';
+import { Bridge, SubstrateDVMBridgeConfig } from '../../../model';
+import { crabConfig } from '../../../config/network';
+
+const crabCrabDVMConfig: SubstrateDVMBridgeConfig = {
+  specVersion: 1180,
+  api: {
+    subql: 'https://api.subquery.network/sq/darwinia-network/wormhole-',
+  },
+};
+
+/**
+ * smart app: crab
+ */
+export const crabCrabDVM = new Bridge<SubstrateDVMBridgeConfig>(
+  omit(crabConfig, 'dvm'),
+  crabConfig,
+  crabCrabDVMConfig,
+  { activeAssistantConnection: true }
+);
+
+const pangolinPangolinDVMConfig: SubstrateDVMBridgeConfig = {
+  specVersion: 27020,
+  api: {
+    subql: 'https://api.subquery.network/sq/darwinia-network/wormhole-',
+  },
+};
+
+/**
+ * smart app: testnet
+ */
+export const pangolinPangolinDVM = new Bridge<SubstrateDVMBridgeConfig>(
+  omit(pangolinConfig, 'dvm'),
+  pangolinConfig,
+  pangolinPangolinDVMConfig,
+  { activeAssistantConnection: true }
+);
