+++ src/bridges/substrate-dvm/config/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from './bridge';
