+++ src/bridges/substrate-substrateDVM/config/bridge.ts
@@ -1,0 +1,34 @@
+import { crabConfig, darwiniaConfig, pangolinConfig, pangoroConfig } from '../../../config/network';
+import { EVOLUTION_DOMAIN } from '../../../config/api';
+import { Bridge, SubstrateSubstrateDVMBridgeConfig } from '../../../model';
+
+const darwiniaCrabDVMConfig: SubstrateSubstrateDVMBridgeConfig = {
+  specVersion: 1180,
+  api: {
+    dapp: 'https://api.darwinia.network',
+    evolution: EVOLUTION_DOMAIN.product,
+    subGraph: 'https://crab-thegraph.darwinia.network/subgraphs/name/wormhole/Sub2SubMappingTokenFactory',
+    subql: 'https://api.subquery.network/sq/darwinia-network/wormhole-',
+    //     subqlMMr: 'https://api.subquery.network/sq/darwinia-network/darwinia-mmr',
+  },
+};
+
+/**
+ * substrate <-> substrate dvm
+ */
+export const darwiniaCrabDVM = new Bridge(darwiniaConfig, crabConfig, darwiniaCrabDVMConfig);
+
+const pangoroPangolinDVMConfig: SubstrateSubstrateDVMBridgeConfig = {
+  specVersion: 27020,
+  api: {
+    dapp: 'https://api.darwinia.network.l2me.com',
+    evolution: EVOLUTION_DOMAIN.dev,
+    subGraph: 'https://pangolin-thegraph.darwinia.network/subgraphs/name/wormhole/Sub2SubMappingTokenFactory',
+    subql: 'https://api.subquery.network/sq/darwinia-network/wormhole-',
+  },
+};
+
+/**
+ * substrate <-> substrate dvm testnet
+ */
+export const pangoroPangolinDVM = new Bridge(pangoroConfig, pangolinConfig, pangoroPangolinDVMConfig);
