--- src/config/bridge/bridges.ts
@@ -1,19 +1,0 @@
-import { crabCrabDVM } from './crab-crabDVM';
-import { darwiniaCrabDVM } from './darwinia-crabDVM';
-import { ethereumCrabDVM } from './ethereum-crabDVM';
-import { ethereumDarwinia } from './ethereum-darwinia';
-import { pangolinPangolinDVM } from './pangolin-pangolinDVM';
-import { pangoroPangolinDVM } from './pangoro-pangolinDVM';
-import { ropstenPangolin } from './ropsten-pangolin';
-import { ropstenPangolinDVM } from './ropsten-pangolinDVM';
-
-export const BRIDGES = [
-  crabCrabDVM,
-  darwiniaCrabDVM,
-  ethereumCrabDVM,
-  ethereumDarwinia,
-  pangolinPangolinDVM,
-  pangoroPangolinDVM,
-  ropstenPangolin,
-  ropstenPangolinDVM,
-];
