--- src/config/bridge/index.ts
@@ -1,9 +1,0 @@
-export * from './bridges';
-export * from './crab-crabDVM';
-export * from './darwinia-crabDVM';
-export * from './ethereum-crabDVM';
-export * from './ethereum-darwinia';
-export * from './pangolin-pangolinDVM';
-export * from './pangoro-pangolinDVM';
-export * from './ropsten-pangolin';
-export * from './ropsten-pangolinDVM';
