--- src/config/erc20.ts
@@ -1,5 +1,0 @@
-export enum RegisterStatus {
-  unregister,
-  registered,
-  registering,
-}
