+++ src/config/index.ts
@@ -1,8 +1,7 @@
 export * from './abi';
 export * from './api';
-export * from './bridge';
+export * from './bridges';
 export * from './constant';
-export * from './erc20';
 export * from './i18n';
 export * from './network';
 export * from './query';
